refactor(context-menu): H.7 — extract startRadio/startInstantMix/downloadAlbum/copyShareLink actions
